requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
pandas==2.2.2
pyarrow==17.0.0
pgvector==0.3.3

